chore(stories): migrate stories to typescript
